requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pdfplumber==0.9.0
scholarly==1.7.9
pandas==2.0.2
openpyxl==3.1.2
xlsxwriter==3.1.2